添加多并发注解支持。

+++ commons/src/main/java/com/github/dactiv/framework/commons/TimeProperties.java
@@ -160,7 +160,7 @@ public class TimeProperties implements Serializable {
      *
      * @return 时间配置
      */
-    public static TimeProperties of(Integer value, TimeUnit unit) {
+    public static TimeProperties of(long value, TimeUnit unit) {
         return new TimeProperties(value, unit);
     }
 
@@ -182,7 +182,7 @@ public class TimeProperties implements Serializable {
      *
      * @return 时间配置
      */
-    public static TimeProperties ofDay(Integer value) {
+    public static TimeProperties ofDay(int value) {
         return new TimeProperties(value, TimeUnit.DAYS);
     }
 
@@ -193,7 +193,7 @@ public class TimeProperties implements Serializable {
      *
      * @return 时间配置
      */
-    public static TimeProperties ofHours(Integer value) {
+    public static TimeProperties ofHours(int value) {
         return new TimeProperties(value, TimeUnit.HOURS);
     }
 
@@ -204,7 +204,7 @@ public class TimeProperties implements Serializable {
      *
      * @return 时间配置
      */
-    public static TimeProperties ofMinutes(Integer value) {
+    public static TimeProperties ofMinutes(int value) {
         return new TimeProperties(value, TimeUnit.MINUTES);
     }
 
@@ -215,7 +215,7 @@ public class TimeProperties implements Serializable {
      *
      * @return 时间配置
      */
-    public static TimeProperties ofSeconds(Integer value) {
+    public static TimeProperties ofSeconds(int value) {
         return new TimeProperties(value, TimeUnit.SECONDS);
     }
 
@@ -226,7 +226,7 @@ public class TimeProperties implements Serializable {
      *
      * @return 时间配置
      */
-    public static TimeProperties ofMilliseconds(Integer value) {
+    public static TimeProperties ofMilliseconds(long value) {
         return new TimeProperties(value, TimeUnit.MILLISECONDS);
     }
 
@@ -237,7 +237,7 @@ public class TimeProperties implements Serializable {
      *
      * @return 时间配置
      */
-    public static TimeProperties ofMicroseconds(Integer value) {
+    public static TimeProperties ofMicroseconds(long value) {
         return new TimeProperties(value, TimeUnit.MICROSECONDS);
     }
 
@@ -248,7 +248,7 @@ public class TimeProperties implements Serializable {
      *
      * @return 时间配置
      */
-    public static TimeProperties ofNanoseconds(Integer value) {
+    public static TimeProperties ofNanoseconds(long value) {
         return new TimeProperties(value, TimeUnit.NANOSECONDS);
     }
